requirements copy.txt
# -------------------------
# Core
# -------------------------
fastapi==0.115.0
uvicorn[standard]==0.30.6

# -------------------------
# Database & ORM
# -------------------------
sqlalchemy==2.0.34
alembic==1.13.2
databases[postgresql]==0.9.0
psycopg2-binary

# -------------------------
# Auth & Security
# -------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-users==13.0.0
fastapi-limiter==0.1.6
slowapi==0.1.9
fastapi-csrf-protect==0.3.4
pydantic==2.5.1  # The core library that causes conflicts
fastapi-mail==1.4.1
email-validator==2.1.0.post1

# -------------------------
# API Utilities
# -------------------------
fastapi-utils==0.7.0  # Note: This library is no longer actively maintained and has limited pydantic v2 support. Consider alternatives.
fastapi-versioning==0.10.0
fastapi-pagination==0.12.27
httpx==0.27.2

# -------------------------
# Logging & Monitoring
# -------------------------
loguru==0.7.2
sentry-sdk==2.13.0
prometheus-fastapi-instrumentator==7.0.0

# -------------------------
# File & Media
# -------------------------
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.4.0
boto3==1.35.0

# -------------------------
# Background Tasks & Jobs
# -------------------------
celery[redis]==5.4.0
apscheduler==3.10.4

# -------------------------
# Dev Tools
# -------------------------
pytest==8.3.3
pytest-asyncio==0.23.8
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.2

# -------------------------
# Email verification helpers (added)
# -------------------------
dnspython>=2.2.0,<3.0
python-dotenv>=1.0.0,<2.0